Move utility function parseHttpHeaders in core/Utils.js
